Use system timestamp since the unix epoch
Previously, we used System.nanoTime() to capture system's timestamp in
nanoseconds, which works fine. However, since this method returns the
timestamp since the machine's boot, it might not be precise for aligning
the trace time with the current actual time. We now capture the offset
of System.nanoTime() with the actual system's time, and after that,
every captured timestamp will sum-up with the offset.

Signed-off-by: Kaveh Shahedi <kaveh.shahedi@ericsson.com>