[test] restore SimpleFormatter.format property modified in test
Fixes: #54

In test file LoggerWithHandlerTest.java, the following property is
modified:

"java.util.logging.SimpleFormatter.format"

Depending on tests order execution, the modified property can
interfere with other tests that do not expect the SimpleFormatter
to have a different format vs default.

This commit takes care to reset the original value of the property,
so that other tests that expect it to have the default value will
have it.

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>