[tests] cleanup temp files created by tests
Fixes #58

Each time the tests are run, several files are created in
the system's temp directory (/tmp on Linux), each time with
new names. Some of them are quite big (from the benchmark
tests) and if one runs the tests many times, it will take-up
some considerable space. (almost 600MB per run)

In this commit, every place in the tests where a temporary
file is created, "deleteOnExit()" is called on the returned
File handle. This makes it so that the JVM will delete each
of these files, before terminating its execution.

Note: I noticed that some files remain in my temp folder,
with extension ".lck" that are created along-sides the test's
temp files, and I think used by the locking mechanism to
insure ordered writing in the files. These are not directly
created by the tests as far as I can see, and they have a size
of zero, and so should not take much disk space.

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>