[example] Add Instrumented LRU cache example
This toy example was inspired by a real-life investigation,
regarding a web-accessible application that had configurable
caches, including a specific one whose performance came under
scrutiny. That investigation produces some interesting traces,
which we aim to approximately replicate with this example.

Real life investigation - some info about that cache:

type: LRU (least recently used) - when the cache is full, the
cache element that was the least recently used is evicted to
make space for a new element to be inserted in its place.

configurable cache size: The number of elements it can hold
is configurable by the app's admins

cache / cached data usage: The cache is used in a special way.
Thought the cache-backed data can be accessed randomly, it's
often accessed sequentially, from first element to the last, in
order. The app is multi-threaded and often multiple reader
threads would be reading the data through the cache, in parallel.

The app provides cache statistics, which seemed quite good
(a high cache "hit" rate was reported).

Investigation:
The cache was configured to be slightly under-sized vs the
data it was caching, which size (number of elements) was
slowly but surely increasing. Given the good reported cache
stats, was it an issue that it was a bit undersized (by ~15%)?
Was the cache, as configured, potentially affecting performance?

Tracing to the rescue!
The cache and reader thread code were instrumented. From the
generated traces, it was possible to distinguish cache hits
vs misses. Capturing traces for a few scenarios, some surprising
details started to emerge!

See the example's README for more details.

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>